made wofi toggle on button press

+++ hyprland/run_wofi.sh
@@ -1,2 +1,5 @@
-pkill wofi
-wofi --config ~/.config/wofi/config/config --style ~/.config/wofi/src/mocha/style.css --show drun -a
+if pgrep -x "wofi" > /dev/null; then
+    pkill wofi
+else
+    wofi --config ~/.config/wofi/config/config --style ~/.config/wofi/src/mocha/style.css --show drun -a
+fi
